requirements.txt
openai
pdfplumber
python-docx
python-dotenv
firebase-admin
google-auth-oauthlib
google-auth
pandas






